New project approaches. Global ordering worked better
Two new folders with different approaches. The global ordering works better than the first one with even/odd forks. I try to create exceptions in my first approach but every time I get huge variations on odd forks or a helgrind issue. This is the trade off and is normal. So I changed for the global ordering and is better overall. Going to refactor and norminette now.